requirements.txt
scikit-learn==1.4.1.post1
numpy==1.26.4
pillow==10.0.1
gunicorn==22.0.0
